requirements.txt
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.22
langchain-text-splitters==0.2.2
langchain-chroma
langchain-huggingface
langchain-text-splitters==0.2.2
transformers==4.41.2         
torch             

python-dotenv==1.0.1
fastapi==0.111.1
langserve==0.2.2
sse_starlette==2.1.2
sentencepiece
sacremoses
protobuf
httpx==0.26.0
bs4==0.0.2
faiss-cpu==1.8.0.post1
tqdm==4.66.4
tiktoken==0.7.0
langchainhub==0.1.20
langgraph==0.1.8
tavily-python==0.3.4
pytest==8.2.2
langchain_google_genai
